+++ CHANGELOG.md
@@ -11,7 +11,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ### Changed
 
-- Stop calling `typeLoader` for built-in scalar names (`String`, `Int`, `Float`, `Boolean`, `ID`), use `types` config for per-schema scalar overrides https://github.com/webonyx/graphql-php/issues/1874
+- Exclusively support `types` for per-schema scalar overrides, not `typeLoader` https://github.com/webonyx/graphql-php/pull/1884
 
 ## v15.31.1